requirements.txt
# Core dependencies
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.21.0
python-dotenv>=0.19.0

# Aliyun SLS SDK
aliyun-log-python-sdk>=0.7.0

# Aliyun sdk core
aliyun-python-sdk-core>=2.16.0

# Aliyun sdk sts
aliyun-python-sdk-sts>=3.1.3

# Aliyun CMS SDK
alibabacloud-cms20240330>=1.0.0
alibabacloud-tea-openapi>=0.3.0
alibabacloud-tea-util>=0.3.0
alibabacloud-credentials>=0.3.0
alibabacloud-sts20150401>=1.1.0

# Jupyter notebook dependencies
jupyter>=1.0.0
ipykernel>=6.0.0
